EmailSubscribeSnippet: Don't use 'no mail' code when subscribing
In my database, the first mail code is 0, where gmc_mail_cause_target
is 'Never mail'. I assume this is not the correct code to use when
subscribing, as this code is also used when unsubscribing.

This change ensures we use the second mail code (which has id 100 and
gmc_mail_cause_target 'Mail').