subscribe: Proper participate route config
Split in two sections, where the one with the subscribe routes lives oin
the authenticated section, and the one without the unsubscribe routes
lives in the unauthenticated section.
Also removed the defaultPostroutes from the post routes, as we don;t
have them here, and remove basePrivilege: false from the authenticated
routes.