github actions: Add phpstan as job
We could also add this as a separate github action if that's preferred.
Now, we'll get a single report for the 'Test suite' action, where you
can see the result of the separate jobs when looking at the details of
the action.

This currently gives me one error that could not be put in the baseline
file.