Order seeds
Ensure the seeds returned by getInfo() are properly ordered, and update
the test to ensure the order is correct.

I found out that this test passes:

    $a = [
        'a' => 1,
        'b' => 2,
    ];
    $b = [
        'b' => 2,
        'a' => 1,
    ];
    $this->assertEquals($a, $b);

Which can be argued for, as both arrays have an key 'a' with value 1
and a key 'b' with value 2. I didn't expect these to be equal, but
according to phpunit (10.3-dev) they are.