[02/09] ratelimit: Configuration
Use a flat config structure to keep things tidy and readable.

This will need to be extended to include other routes to which the rate
limiter needs to be applied.

If custom middleware lists are configured in the Route config, then care
must be taken that the RatelimitMiddleware class is included for rate
limiting to actually work.