[04/09] ratelimit: Add RatelimitMiddleware to default middleware
This may need some optimization. Currently the order of several middleware
classes is different between the authenticated and unauthenticated
routes. Also, for the rate limiter to have the best effect it should be
added as early as possible, but after the ClientIpMiddleware and
authentication middlewares.

Also, several authentication routes have their own middleware config, so
the rate limiter does not apply to them yet.