Fix for Check all rounds function
The 'Check all rounds' function didn't complete on the test hosts,
because prevRoundId happened to be null, which caused a backtrace.
Since AnyStepEngine already allowed prevRoundId to be null, I allowed
this in the other StepEngine classes as well.
Also, don't try to assign if relationId is null.

Not sure if this should be a hard error instead.