[28/36] PatchRepository: Allow empty migrations
For dynamic migrations, I used SELECT 1 or DO 0 to try and get a noop
statement, but that breaks things silently somewhere causing the
migration not to be added to the gems__migration_logs table, which means
it remains in the to-do list.
Fix that by allowing empty migrations, so we no longer need to work with
dummy statements.