RoundModel: Break self referencing foreign key before deleting
The gro_valid_for_id column usually contains the id of the row itself.
Since it also has a foreign key, this makes it impossible to delete such
a row. In order to fix this, we first set the value of gro_valid_for_id
to null before deleting the row.