EditTrackTokenSnippet: Differ between changed and recalculated tokens
This fixes the issue that when a token was changed in such a way that
no recalculations were neccesary (such as changing a manual date), the
user would get the message 'No changes to save!'. Now we make a
distinction between the number of changed tokens (the current one) and
the number of recalculated tokens, which has its own message to the
user.