HelperAdapter: Don't return invalidated cache items
While the cachedResultFetcher worked properly, entries retrieved from
the cache directly gave cache hits even after the entries were
invalidated. The reason was that the cachedResultFetcher called

    $this->cache->hasItem($cacheKey)

which returns false for invalidated keys.
The HelperAdapter checked if the items pool had the cache item
(bypassing the tag validation) and returned it if true.

As optimization, instead of doing:
    if ($this->cache->hasItem($cacheKey)) {
        return $this->cache->getCacheItem($cacheKey);
    }
where the item has to be requested twice (once to see if the get was a
hit, and once to return the result) we get it only once, and return the
result if it was a hit.