[2/6] Simplify route matching
This fixes an issue where some routes (such as those specified with
'seeds') were never matched, because the end of string match did not
allow any dots between the pattern and the end of the string, but the
route names do contain dots (setup.database.patches.show).

We drop the ability to specify patterns here, which means we have to
include all the respondent.something.show routes explicitly.