Ensure respondents can only be added to accessible orgs that accept new respondents
This ensures that it is not possible to add respondents to an
organization that doesn't have the gor_add_respondents setting, or to an
organization that the user doesn't have access to.

If the current organization of the user does not accept new patients, we
default to the first organization that does (although it might be better
to just present a dropdown in that case).

If the user doesn't have access to any organizations that accept new
respondents, we throw an exception.