Fix reception code filter for unanswered deletion codes
The three functions getTokenRestoreCodes(),
getCompletedTokenDeletionCodes() and getUnansweredTokenDeletionCodes()
are all three called from the getReceptionCodes() function in the
DeleteTrackTokenSnippet.

It seems plausible that these all deal with the same type of codes, so I
think the filter for SURVEY_TYPE_FIELD should be the same for all three.

Note that one filters on SURVEY_TYPE_FIELD == 1, and two filter on
SURVEY_TYPE_FIELD != 0. Not sure if this field can have other values
than 0 and 1.