Update search filter to include data from all accessible organizations
This does two things: on the tracks page it lists all tracks available
to the current organization or to an organization where the current
organization has access to.
It also ensures that a track that exists for an accessible organization
is visible from the context of the current organizations.

There's still some work to do though:
  - Showing a track of an accessible organization switches the context
    to that of the accessible organization (e.g. the organization
    parameter in the URL is not the current, but the accessable org)
  - Editing a track of an accessible organization from the context
    of the current organization changes the value of
    gr2t_id_organization for that track to the current org, effectively
    moving it, which is not what we want.