Calculate date from start of the week
When looking through this code I found an inconsistency. When
calculating the month or year intervals, the code starts at the
beginning of the interval (e.g. "first day of this month", "first day of
January this year"). However, for the week interval it used "this
monday", which translates to 'the first upcoming monday' on my system.

So I think this maybe needs to be "monday this week" to get the same
behaviour?