Preserve flash messages in logout handler
This fixes a problem which happens in the Embed Login functionality, and
possibly in other cases as well. If, during embed login, the patient is
not found, a flash message is set, and a redirect to the logout route is
issued. The logout handler clears the session and redirects to the login
page. The flash messages are therefore never shown.

Since there is no way to get all messages with the correct message
type, this unfortunately is a bit convoluted.