Init answerFilter property as null
The code tests if this property contains an instance of
AnswerNameFilterInterface, but if the doctype declares that it cannot be
anything else, then phpstan complains about the check being useless.
If we declare it as null, we avoid this complaint and we also avoid a
potential problem when we access the property before it has a value.

Currently not used anywhere, as far as I can see.