[3/3] Don't set gr2o_opened to current time for BrowseSearchAction
This partially fixes a weird problem. When a user logs in, the search
filter contains the 'id2' key, probably because the organization select
is part of the login form. That in turn causes the gr2o_opened key to be
shown (see RespondentTableSnippet::addBrowseColumn1()).

When the respondent model is created, immediately the edit action is
applied to it, which adds the gr2o_opened key with current time.
If we than run the BrowseSearchAction (using the same model object), the
gr2o_opened key is already set.

That means that immediately after logging in, the user sees all
respondents in the browse screen with the same opened time (the current
time).

After clicking the navigation controls on the bottom of the form, the
'id2' key is removed from the search filter and the
RespondentTableSnippet shows the organization instead.

This doesn't fix the switch from gr2o_opened to organization name, but
does ensure that the correct value of gr2o_opened is shown.