[1/3] Sanitize the export ID
If a questionnaire has a name that includes encodable characters (such
as '&') then the URL encoding that happens when this ID is used as an
URL parameter causes it to be encoded as %26amp%3B (e.g. double
encoding, from & to &amp to %26amp%3B). This causes the download to fail.

Since we create an ID from user input, let's be safe and ensure the ID
never contains any unwanted characters.